requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cachetools==6.2.4
cbor2==5.9.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.6
daphne==4.2.1
dj-database-url==3.1.0
Django==5.2.12  
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
google-api-core==2.29.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==26.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.4.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
zope.interface==8.2